removed Users class, distribute methods
