+++ lib/check_game.rb
@@ -1,9 +1,4 @@
 class GameCheck
-  def vallidate_position(data)
-    pattern = /[0-8]/
-    data.length == 1 && data.match?(pattern) ? true : false
-  end
-
   def vallidate_name(name)
     pattern = /^[a-zA-Z]*$/
     valid_name = name.match? pattern
@@ -12,5 +7,10 @@ class GameCheck
 
   def vallidate_symbol(symbol)
     symbol.eql?('X') || symbol.eql?('Y') ? true : false
+  end
+
+  def toggle_symbol(parameter)
+    return 'Y' if parameter == 'X'
+    return 'X' if parameter == 'Y'
   end
 end
