--- lib/users.rb
@@ -1,27 +1,0 @@
-require_relative '../lib/check_game'
-require_relative '../lib/board'
-class Users
-  def select_symbol
-    return first_symbol if vallid_symbol == true
-  end
-
-  def assign_symbol(parameter, player_one, player_two)
-    if parameter == 'X'
-      second_symbol = 'Y'
-      "#{player_one} selected #{parameter} and #{player_two} now takes #{second_symbol}"
-    elsif parameter == 'Y'
-      second_symbol = 'X'
-      "#{player_one} selected #{parameter} and #{player_two} now takes #{second_symbol}"
-    end
-  end
-
-  def toggle_symbol(parameter)
-    return 'Y' if parameter == 'X'
-    return 'X' if parameter == 'Y'
-  end
-
-  def play(player_choose, vallid_symbol)
-    board = GameBoard.new
-    board.place_symbol(player_choose, vallid_symbol) if player_choose.is_a?(Numeric)
-  end
-end
